Move action naming and description into an injectable ActionNamingProvider. Have maps override this rather than implement an option ViewSettings method.
