+++ src/maps/disco/view_settings.ts
@@ -1,7 +1,6 @@
 import { ClickTarget, OnClickRegister } from "../../client/grid/click_target";
 import { useAction } from "../../client/services/action";
 import { useGrid, useInjected } from "../../client/utils/injection_context";
-import { Action } from "../../engine/state/action";
 import { MapViewSettings } from "../view_settings";
 import { DiscoMoveHelper } from "./deliver";
 import { ProductionAction } from "./production";
@@ -13,13 +12,6 @@ export class DiscoInfernoViewSettings
   implements MapViewSettings
 {
   getMapRules = DiscoInfernoRules;
-
-  getActionDescription(action: Action): string | undefined {
-    if (action === Action.PRODUCTION) {
-      return "Draw two cubes and place them in one city after the move goods step.";
-    }
-    return undefined;
-  }
 
   moveGoodsMessage(): string | undefined {
     const helper = useInjected(DiscoMoveHelper);
