+++ src/maps/germany/view_settings.ts
@@ -1,4 +1,3 @@
-import { Action } from "../../engine/state/action";
 import { GermanyRules } from "./rules";
 import { GermanyMapSettings } from "./settings";
 import { MapViewSettings } from "../view_settings";
@@ -10,11 +9,4 @@ export class GermanyViewSettings
 {
   getMapRules = GermanyRules;
   getTexturesLayer = GermanyRivers;
-
-  getActionDescription(action: Action): string | undefined {
-    if (action === Action.ENGINEER) {
-      return "Build one tile (the most expensive one) at half price (rounded down).";
-    }
-    return undefined;
-  }
 }
