+++ src/maps/madagascar/view_settings.ts
@@ -1,4 +1,3 @@
-import { Action } from "../../engine/state/action";
 import { MapViewSettings } from "../view_settings";
 import { getActionCaption } from "./action_caption";
 import { MadagascarRules } from "./rules";
@@ -11,14 +10,4 @@ export class MadagascarViewSettings
   getMapRules = MadagascarRules;
 
   getActionCaption = getActionCaption;
-
-  getActionDescription(action: Action): string | undefined {
-    if (action === Action.LOCOMOTIVE) {
-      return "Immediately, increase your locomotive by one, but you cannot build track this turn.";
-    }
-    if (action === Action.URBANIZATION) {
-      return "Place a new city on any town during the build step, but may only build one track tile.";
-    }
-    return undefined;
-  }
 }
