+++ src/maps/scotland/view_settings.ts
@@ -1,4 +1,3 @@
-import { Action } from "../../engine/state/action";
 import { MapViewSettings } from "../view_settings";
 import { ScotlandRules } from "./rules";
 import { ScotlandRivers } from "./rivers";
@@ -10,11 +9,4 @@ export class ScotlandViewSettings
 {
   getMapRules = ScotlandRules;
   getTexturesLayer = ScotlandRivers;
-
-  getActionDescription(action: Action): string | undefined {
-    if (action === Action.TURN_ORDER_PASS) {
-      return "Become first player in the next round skipping auction phase.";
-    }
-    return undefined;
-  }
 }
