+++ src/maps/st-lucia/view_settings.ts
@@ -1,4 +1,3 @@
-import { Action } from "../../engine/state/action";
 import { MapViewSettings } from "../view_settings";
 import { StLuciaRivers } from "./rivers";
 import { StLuciaRules } from "./rules";
@@ -11,10 +10,4 @@ export class StLuciaViewSettings
   getTexturesLayer = StLuciaRivers;
 
   getMapRules = StLuciaRules;
-
-  getActionDescription?(action: Action): string | undefined {
-    if (action === Action.TURN_ORDER_PASS) {
-      return "Go first in the next auction.";
-    }
-  }
 }
